refactor: 사용자 API 기능별로 분리
