+++ app/src/main/java/com/techcourse/controller/AuthController.java
@@ -13,14 +13,25 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class LoginController {
+public class AuthController {
 
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+    private static final Logger log = LoggerFactory.getLogger(AuthController.class);
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView showLoginView(final HttpServletRequest req, final HttpServletResponse res) {
+        final String forwardUri = UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return "redirect:/index.jsp";
+                })
+                .orElse("/login.jsp");
+        return new ModelAndView(new JspView(forwardUri));
+    }
 
     @RequestMapping(value = "/login", method = RequestMethod.POST)
     public ModelAndView login(final HttpServletRequest req, final HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
-            return new ModelAndView(new JspView("redirect:/index.jsp"));
+            return new ModelAndView(new JspView("redirect:/"));
         }
 
         final String forwardUri = InMemoryUserRepository.findByAccount(req.getParameter("account"))
@@ -33,11 +44,18 @@ public class LoginController {
         return new ModelAndView(new JspView(forwardUri));
     }
 
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) {
+        final var session = req.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return new ModelAndView(new JspView("redirect:/"));
+    }
+
     private String login(final HttpServletRequest request, final User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             final var session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
+            return "redirect:/";
         } else {
             return "redirect:/401.jsp";
         }
